+++ packages/webapp/src/single/MediaNav.tsx
@@ -4,16 +4,18 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import * as icons from '@fortawesome/free-solid-svg-icons'
 
 import { useSearchStore } from "../store/search-store";
-import { useAppConfig } from "../utils/useAppConfig";
+import { useAppConfig } from "../config/useAppConfig";
 
 import { getHigherPreviewUrl, getLowerPreviewUrl } from '../utils/preview'
 import { usePreviewSize } from "./usePreviewSize";
 import { classNames } from '../utils/class-names'
+import { MediaViewDisableFlags } from "./MediaViewPage";
 
 export const MediaNav = ({current, prev, next, listLocation, showNavigation, dispatch}) => {
   const query = useSearchStore(state => state.query);
   const previewSize = usePreviewSize()
   const appConfig = useAppConfig()
+  const diabledFlags = appConfig.pages?.mediaView?.disabled || [] as MediaViewDisableFlags
 
   const loadImage = async (url: string | false) => {
     return new Promise((resolve) => {
@@ -57,20 +59,20 @@ export const MediaNav = ({current, prev, next, listLocation, showNavigation, dis
   return (
     <>
       <div className={classNames('absolute z-10 top-4 right-4 flex gap-2')}>
-        {!appConfig.removedViewerStream &&
+        {!diabledFlags?.includes('nav') && 
           <a onClick={() => dispatch({type: 'list'})} className={classNames(buttonClass, itemClass, 'bg-transparent hover:bg-gray-400/40')} title="Show media stream (ESC)">
             <FontAwesomeIcon icon={icons.faXmark} className={iconClass}/>
           </a>
         }
       </div>
-      {!appConfig.removedViewerNav && prev &&
+      {!diabledFlags?.includes('nav') && prev &&
         <div className={classNames('absolute z-10 left-4 top-1/2 -translate-y-1/2', itemClass)}>
           <a onClick={() => dispatch({type: 'prev'})} className={classNames(buttonClass, buttonBgClass)} title="Show previous media (left arrow)">
             <FontAwesomeIcon icon={icons.faChevronLeft} className={iconClass}/>
           </a>
         </div>
       }
-      {!appConfig.removedViewerNav && next &&
+      {!diabledFlags?.includes('nav') && next &&
         <div className={classNames('absolute z-10 right-4 top-1/2 -translate-y-1/2', itemClass)}>
           <a onClick={() => dispatch({type: 'next'})} className={classNames(buttonClass, buttonBgClass)} title="Show next media (right arrow)">
             <FontAwesomeIcon icon={icons.faChevronRight} className={iconClass}/>
@@ -78,32 +80,32 @@ export const MediaNav = ({current, prev, next, listLocation, showNavigation, dis
         </div>
       }
       <div className={classNames('absolute z-10 bottom-4 left-1/2 -translate-x-1/2 flex gap-2')}>
-        {!appConfig.removedViewerStream && listLocation &&
+        {!diabledFlags.includes('nav') && listLocation &&
           <a onClick={() => dispatch({type: 'list'})} className={classNames(buttonClass, buttonBgClass, itemClass)} title="Show media stream (ESC)">
             <FontAwesomeIcon icon={icons.faTh} className={iconClass}/>
           </a>
         }
-        {!appConfig.removedViewerMap && hasGeo &&
+        {!diabledFlags.includes('map') && hasGeo &&
           <a onClick={() => dispatch({type: 'map'})} className={classNames(buttonClass, buttonBgClass, itemClass)} title="Show map of entry (m)">
             <FontAwesomeIcon icon={icons.faMap} className={iconClass}/>
           </a>
         }
-        {!appConfig.removedViewerLeaf && current?.similarityHash &&
+        {!diabledFlags.includes('similar') && current?.similarityHash &&
           <a onClick={() => dispatch({type: 'similar'})} className={classNames(buttonClass, buttonBgClass, itemClass)} title="Show similar images (s)">
             <FontAwesomeIcon icon={icons.faSeedling} className={iconClass}/>
           </a>
         }
-        {!appConfig.removedViewerYears && query.type != 'none' &&
+        {!appConfig.pages?.disabled?.includes('date') && query.type != 'none' &&
           <a onClick={() => dispatch({type: 'chronology'})} className={classNames(buttonClass, buttonBgClass, itemClass)} title="Show chronology (c)">
             <FontAwesomeIcon icon={icons.faClock} className={iconClass}/>
           </a>
         }
-        {!appConfig.removedViewerAI && current && (current.faces?.length > 0 || current.objects?.length > 0) &&
+        {!diabledFlags.includes('annotation') && current && (current.faces?.length > 0 || current.objects?.length > 0) &&
           <a onClick={() => dispatch({type: 'toggleAnnotations'})} className={classNames(buttonClass, buttonBgClass, itemClass)} title="Show object and face annotations (a)">
             <FontAwesomeIcon icon={icons.faUsersViewfinder} className={iconClass}/>
           </a>
         }
-        {!appConfig.removedViewerInfo && current &&
+        {!diabledFlags.includes('detail') && current &&
           <a onClick={() => dispatch({type: 'toggleDetails'})} className={classNames(buttonClass, buttonBgClass, itemClass)} title="Show detail info (i)">
             <FontAwesomeIcon icon={icons.faInfo} className={iconClass}/>
           </a>