+++ packages/webapp/src/single/MediaView.tsx
@@ -7,7 +7,9 @@ import {
 } from "react-router-dom";
 import Hammer from 'hammerjs';
 import { useHotkeys } from 'react-hotkeys-hook';
-import { useAppConfig } from "../utils/useAppConfig";
+import Logger from '@home-gallery/logger'
+
+import { useAppConfig } from "../config/useAppConfig";
 import { useEntryStore } from "../store/entry-store";
 import { useSearchStore } from "../store/search-store";
 import { useSingleViewStore } from "../store/single-view-store";
@@ -22,6 +24,10 @@ import { Zoomable } from "./Zoomable";
 import useBodyDimensions from "../utils/useBodyDimensions";
 import { classNames } from '../utils/class-names'
 import { SingleTagDialogProvider } from "../dialog/tag-dialog-provider";
+import { useMediaViewHotkeys } from "./useMediaViewHotkeys";
+import { MediaViewDisableFlags } from "./MediaViewPage";
+
+const log = Logger('MediaView')
 
 const findEntryIndex = (location, entries, id) => {
   if (location.state?.index && entries[location.state.index]?.id.startsWith(id)) {
@@ -50,34 +56,9 @@ const scaleDimensions = (media, device) => {
 
 const encodeUrl = (url: string) => url.replace(/[\/]/g, char => encodeURIComponent(char))
 
-const hotkeysToAction = {
-  'home': 'first',
-  'arrowleft,j,backspace': 'prev',
-  'ctrl+arrowleft': 'prev-10',
-  'ctrl+shift+arrowleft': 'prev-100',
-  'arrowright,k,space': 'next',
-  'ctrl+arrowright': 'next-10',
-  'ctrl+shift+arrowright': 'next-100',
-  'end': 'last',
-  'esc': 'list',
-  'i': 'toggleDetails',
-  'a': 'toggleAnnotations',
-  's': 'similar',
-  'c': 'chronology',
-  't': 'toggleNavigation',
-  'm': 'map'
-}
-
-const allHotkeys = Object.keys(hotkeysToAction).join(',')
-const hotkeyToAction = Object.entries(hotkeysToAction).reduce((result, [hotkeys, action]) => {
-  hotkeys.split(',').forEach(hotkey => {
-    result[hotkey] = action
-  })
-  return result
-}, {})
-
 export const MediaView = () => {
   const appConfig = useAppConfig();
+  const disableFlags = appConfig.pages?.mediaView?.disabled || [] as MediaViewDisableFlags
   let { id } = useParams();
   let location = useLocation();
   const navigate = useNavigate();
@@ -97,8 +78,9 @@ export const MediaView = () => {
   const setShowNavigation = useSingleViewStore(actions => actions.setShowNavigation);
 
   const [hideNavigation, setHideNavigation] = useState(false)
-  
-  const [zoomFactor, setZoomFactor] = useState(1);
+  const [zoomFactor, setZoomFactor] = useState(1)
+
+  const [hotkeys, hotkeyToAction] = useMediaViewHotkeys();
 
   let index = findEntryIndex(location, entries, id);
 
@@ -111,8 +93,6 @@ export const MediaView = () => {
   const isUnknown = !current || (['image', 'rawImage', 'video'].indexOf(current.type) < 0)
 
   const key = current ? current.id : (Math.random() * 100000).toFixed(0);
-  const scaleSize = scaleDimensions(current, dimensions);
-  //console.log(scaleSize, dimensions, current);
 
   useEffect(() => { id && setLastId(id) }, [id])
   useEffect(() => { index >= 0 && setLastIndex(index) }, [index])
@@ -133,11 +113,11 @@ export const MediaView = () => {
       const negate = prevNextMatch[1] == 'prev' ? -1 : 1
       const i = Math.min(entries.length - 1, Math.max(0, index + (negate * offset)))
       viewEntry(i)
-    } else if (type === 'similar' && current?.similarityHash) {
+    } else if (type === 'similar' && current?.similarityHash && !disableFlags.includes('annotation')) {
       navigate(`/similar/${current.shortId}`);
-    } else if (type === 'toggleDetails') {
+    } else if (type === 'toggleDetails' && !disableFlags.includes('detail')) {
       setShowDetails(!showDetails);
-    } else if (type === 'toggleAnnotations') {
+    } else if (type === 'toggleAnnotations' && !disableFlags.includes('annotation')) {
       setShowAnnotations(!showAnnotations);
     } else if (type === 'toggleNavigation') {
       setShowNavigation(!showNavigation);
@@ -156,7 +136,7 @@ export const MediaView = () => {
       setHideNavigation(false);
     } else if (type == 'search') {
       navigate(`/search/${encodeUrl(action.query)}`);
-    } else if (type == 'map') {
+    } else if (type == 'map' && current?.latitude && current?.longitude && !disableFlags.includes('map')) {
       navigate(`/map?lat=${current.latitude.toFixed(5)}&lng=${current.longitude.toFixed(5)}&zoom=14`, {state: {listLocation}})
     }
   }
@@ -169,20 +149,17 @@ export const MediaView = () => {
     }
   }
 
-  useHotkeys(allHotkeys, (ev, handler) => {
+  useHotkeys(hotkeys, (ev, handler) => {
     const handlerKey = (handler.ctrl ? 'ctrl+' : '') + (handler.shift ? 'shift+' : '') + (handler.alt ? 'alt+' : '') + (handler.keys || []).join('+')
     const action = hotkeyToAction[handlerKey]
-    if (action) {
-      if ((action === 'prev' || action === 'next' || action.startsWith('prev-') || action.startsWith('next-')) && appConfig.removedViewerNav) return;
-      if (action === 'list' && appConfig.removedViewerStream) return;
-      if (action === 'map' && appConfig.removedViewerMap) return;
-      if (action === 'chronology' && appConfig.removedViewerLeaf) return;
-      if (action === 'similar' && appConfig.removedViewerAI) return;
-      if (action === 'toggleDetails' && appConfig.removedViewerInfo) return;
-      if (action === 'toggleAnnotations' && appConfig.removedViewerAI) return;
-      dispatch({type: action})
-      ev.preventDefault()
+
+    if (!action) {
+      log.warn(`Hotkey action of ${handlerKey} not found`)
+      return
     }
+
+    dispatch({type: action})
+    ev.preventDefault()
   }, [index, showDetails, showAnnotations, showNavigation])
 
   const mediaVanishes = index < 0 && lastIndex >= 0 && entries.length > 0
@@ -194,8 +171,8 @@ export const MediaView = () => {
     dispatch({type: 'list'})
   }
 
-  //console.log('Media object', current, showDetails);
-  //console.log(appConfig)
+  console.log('Media object', current);
+
   return (
     <>
       <SingleTagDialogProvider>
@@ -206,8 +183,8 @@ export const MediaView = () => {
                 <MediaNav index={index} current={current} prev={prev} next={next} listLocation={listLocation} showNavigation={showNavigation} dispatch={dispatch} />
               }
               {isImage &&
-                <Zoomable key={key} childWidth={current.width} childHeight={current.height} onSwipe={onSwipe} onZoom={(scale) => setZoomFactor(scale)}>
-                  <MediaViewImage key={key} media={current} next={next} prev={prev} showAnnotations={showAnnotations} hqZoom={appConfig.HQzoom} zoomFactor={zoomFactor}/>
+                <Zoomable key={key} childWidth={current.width} childHeight={current.height} onSwipe={onSwipe} onZoom={setZoomFactor}>
+                  <MediaViewImage key={key} media={current} next={next} prev={prev} showAnnotations={showAnnotations} zoomFactor={zoomFactor}/>
                 </Zoomable>
               }
               {isVideo &&