+++ packages/webapp/vite.config.ts
@@ -2,6 +2,8 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import tailwindcss from '@tailwindcss/vite'
 import { VitePWA } from 'vite-plugin-pwa'
+import pwaConditionalPlugin from './vite/pwa-conditional-plugin'
+import injectAppState from './vite/inject-app-state'
 
 const proxyTarget = 'http://localhost:3000';
 const serverUrls = [
@@ -27,8 +29,9 @@ export default defineConfig(() => {
       tailwindcss(),
       VitePWA({
         registerType: 'autoUpdate',
-        includeAssets: ['favicon.ico', 'robots.txt', 'logo192.png', 'logo512.png', 'logo.svg'],
+        includeAssets: ['favicon.ico', 'logo192.png', 'logo512.png', 'logo.svg'],
         manifest: {
+          id: '/',
           name: 'HomeGallery App',
           short_name: 'HomeGallery',
           description: 'All personal photos in your pocket',
@@ -60,11 +63,25 @@ export default defineConfig(() => {
               sizes: '512x512'
             }
           ]
+        },
+        workbox: {
+          navigateFallbackDenylist: [/\/api\//, /\/files\//],
+        }
+      }),
+      pwaConditionalPlugin({ disabled: false }),
+      injectAppState({
+        // See gallery.config-example.yml for all available options
+        disabled: true,
+        state: {
+          sources: [
+            // enable downloadable sources
+            //{ indexName: "Camera roll", downloadable: true }
+          ]
         }
       })
     ],
     resolve: {
-      extensions: ['.ts', '.tsx']
+      extensions: ['.ts', '.tsx', '.json']
     },
     server: {
       proxy